requirements.txt
asgiref==3.4.1
Django==4.0
sqlparse==0.4.2
Pillow==8.3.2
pytz==2021.1
djangorestframework~=3.12.4
psycopg2-binary==2.9.2
django-multiselectfield==0.0.2
